+++ Misc/NEWS.d/next/Tests/2018-11-30-17-18-56.bpo-35352.8bD7GC.rst
@@ -1,0 +1,1 @@
+Modify test_asyncio to use the certificate set from the test directory.
